requirements.txt
Flask>=2.0.1
Pillow>=10.0.0
Werkzeug>=2.0.3
numpy>=1.24.0 